Rename InputBar and polish codes
- Rename InputBar.js to Input.js
- Remove unecessary props on Input and TodoList
